Combat changes and bug fixes
Added the combo and proc systems
Added scripts for most weaponskill and spells as well as some abilities and status effects
Added support for multihit attacks
Added AbilityState for abilities
Added hiteffects that change based on an attack's parameters
Added positionals

Changed how targeting works for battlecommands

Fixed bug that occurred when moving or swapping hotbar commands
Fixed bug that occurred when losing status effects

+++ sql/characters_statuseffect.sql
@@ -1,8 +1,8 @@
--- MySQL dump 10.13  Distrib 5.7.10, for Win64 (x86_64)
+-- MySQL dump 10.13  Distrib 5.7.11, for Win64 (x86_64)
 --
--- Host: localhost    Database: ffxiv_database
+-- Host: localhost    Database: ffxiv_server
 -- ------------------------------------------------------
--- Server version	5.7.10-log
+-- Server version	5.7.11
 
 /*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
 /*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
@@ -24,13 +24,13 @@ DROP TABLE IF EXISTS `characters_statuseffect`;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `characters_statuseffect` (
   `characterId` int(10) unsigned NOT NULL,
-  `statusId` smallint(5) unsigned NOT NULL,
+  `statusId` mediumint(8) unsigned NOT NULL,
   `magnitude` bigint(20) unsigned NOT NULL,
   `duration` int(10) unsigned NOT NULL,
   `tick` int(10) unsigned NOT NULL,
   `tier` tinyint(3) unsigned NOT NULL,
   `extra` bigint(20) unsigned NOT NULL,
-  PRIMARY KEY (`characterId`, `statusId`)
+  PRIMARY KEY (`characterId`,`statusId`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
@@ -52,4 +52,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2016-06-07 22:54:49
+-- Dump completed on 2018-02-15  0:04:42
